Add test coverage to flatUnpack API
