+++ subprojects/core-utils/src/test/groovy/dev/nokee/utils/DeferredUtilsTest.groovy
@@ -1,28 +1,12 @@
 package dev.nokee.utils
 
-import kotlin.jvm.functions.Function0
-import org.gradle.api.Named
-import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
-import org.gradle.testfixtures.ProjectBuilder
-import spock.lang.Specification
 import spock.lang.Subject
 import spock.lang.Unroll
 
-import java.util.concurrent.Callable
-import java.util.function.Supplier
-
-import static dev.nokee.utils.DeferredUtils.isDeferred
-import static dev.nokee.utils.DeferredUtils.isNestableDeferred
-import static dev.nokee.utils.DeferredUtils.realize
-import static dev.nokee.utils.DeferredUtils.unpack
+import static dev.nokee.utils.DeferredUtils.*
 
 @Subject(DeferredUtils)
-class DeferredUtilsTest extends Specification {
-	def project = ProjectBuilder.builder().build()
-	def objects = project.objects
-	def providers = project.providers
-
+class DeferredUtilsTest extends DeferredUtils_BaseSpec {
 	def "can unpack provider types"() {
 		expect:
 		unpack(propertyOf('foo')) == 'foo'
@@ -64,7 +48,7 @@ class DeferredUtilsTest extends Specification {
 	@Unroll
 	def "can realize deferred collection"(collectionFactoryMethod) {
 		given:
-		def collection = objects."${collectionFactoryMethod}"(String)
+		def collection = objectFactory."${collectionFactoryMethod}"(String)
 
 		and:
 		def realizedElements = []
@@ -90,7 +74,7 @@ class DeferredUtilsTest extends Specification {
 	@Unroll
 	def "can realize empty deferred collection"(collectionFactoryMethod) {
 		given:
-		def collection = objects."${collectionFactoryMethod}"(String)
+		def collection = objectFactory."${collectionFactoryMethod}"(String)
 
 		when:
 		realize(collection)
@@ -109,69 +93,4 @@ class DeferredUtilsTest extends Specification {
 //		expect:
 //
 //	}
-
-	protected Property<String> propertyOf(String value) {
-		return propertyOf(String, value)
-	}
-
-	protected <T> Property<T> propertyOf(Class<T> type, T value) {
-		return objects.property(type).value(providerOf(value))
-	}
-
-	protected <T> Provider<T> providerOf(T value) {
-		return providers.provider { ThrowIfResolvedGuard.resolve(value) }
-	}
-
-	protected <T> Supplier<T> supplierOf(T value) {
-		return new Supplier<T>() {
-			@Override
-			T get() {
-				return ThrowIfResolvedGuard.resolve(value)
-			}
-		}
-	}
-
-	protected <T> Closure<T> closureOf(T value) {
-		return { ThrowIfResolvedGuard.resolve(value) }
-	}
-
-	protected <T> Function0<T> kotlinFunctionOf(T value) {
-		return new Function0<T>() {
-			@Override
-			T invoke() {
-				return ThrowIfResolvedGuard.resolve(value)
-			}
-		}
-	}
-
-	protected <T> Callable<T> callableOf(T value) {
-		return new Callable<T>() {
-			@Override
-			T call() throws Exception {
-				return ThrowIfResolvedGuard.resolve(value)
-			}
-		}
-	}
-
-	protected static Named namedElement(String name) {
-		return new Named() {
-			@Override
-			String getName() {
-				return name
-			}
-		}
-	}
-
-	protected static ThrowIfResolvedGuard throwIfResolved() {
-		return new ThrowIfResolvedGuard()
-	}
-
-	private static class ThrowIfResolvedGuard {
-		static <T> T resolve(T obj) {
-			if (obj instanceof ThrowIfResolvedGuard) {
-				throw new AssertionError((Object)"Should not resolve")
-			}
-			return obj;
-		}
-	}
 }
